ConfigAgent: move `BounceDatabase(`
